refactor: edit tests, refactoring
